requirements.txt
django==4.0.6
Pillow==9.2.0
django-debug-toolbar==3.5.0
djangorestframework~=3.13.1
drf-spectacular==0.22.1
celery==5.2.7
django-celery-results==2.4.0
redis==4.3.4
django-redis==5.2.0
python-dotenv==0.20.0
gunicorn==20.1.0
django-dotenv==1.4.2
sentry-sdk==1.7.0
django_heroku==0.3.1
dj-database-url==0.5.0
djangorestframework-simplejwt==5.2.0